DOC-142 peters deployment to separate folder
