+++ datasafe-rest-impl/deployment-peter/0.build.jar.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+mv ../src/main/resources/logback-spring.xml ../src/main/resources/weg-logback-spring.xml
+mvn clean package -f ../pom.xml -DskipTests
+mv ../src/main/resources/weg-logback-spring.xml ../src/main/resources/logback-spring.xml
+VERSION=0.0.9-SNAPSHOT
+cp ../target/datasafe-rest-impl-${VERSION}.jar datasafe-rest-server.jar
