+++ datasafe-rest-impl/src/test/java/de/adorsys/datasafe/rest/DatasafeRestApplicationTests.java
@@ -1,6 +1,6 @@
 package de.adorsys.datasafe.rest;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
