+++ datasafe-rest-impl/src/test/java/de/adorsys/datasafe/rest/controller/DocumentControllerTest.java
@@ -1,10 +1,10 @@
 package de.adorsys.datasafe.rest.controller;
 
-import de.adorsys.datasafe.business.api.inbox.InboxServiceImpl;
 import de.adorsys.datasafe.business.impl.service.DefaultDatasafeServices;
+import de.adorsys.datasafe.privatestore.impl.PrivateSpaceServiceImpl;
 import lombok.SneakyThrows;
 import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,11 +26,10 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class InboxControllerTest {
+public class DocumentControllerTest {
 
     private static final String TEST_USER = "test";
     private static final String TEST_PASS = "test";
-    private static final String TEST_PATH = "test.txt";
 
     @Autowired
     MockMvc mvc;
@@ -39,32 +38,21 @@ public class InboxControllerTest {
     DefaultDatasafeServices dataSafeService;
 
     @MockBean
-    private InboxServiceImpl inboxService;
+    private PrivateSpaceServiceImpl privateSpaceService;
 
     @Before
     public void setup() {
         MockitoAnnotations.initMocks(this);
-        when(dataSafeService.inboxService()).thenReturn(inboxService);
+        when(dataSafeService.privateService()).thenReturn(privateSpaceService);
     }
 
     @SneakyThrows
     @Test
-    public void sendDocumentToInboxTest() {
-        when(dataSafeService.inboxService().write(any())).thenReturn(new ByteArrayOutputStream());
+    public void readDocumentTest() {
+        when(dataSafeService.privateService().read(any())).thenReturn(new ByteArrayInputStream("hello".getBytes()));
 
-        mvc.perform(put("/inbox/{path}", TEST_PATH)
-                .contentType(MediaType.APPLICATION_OCTET_STREAM_VALUE)
-                .header("user", TEST_USER)
-        )
-                .andExpect(status().isOk());
-    }
-
-    @SneakyThrows
-    @Test
-    public void readFromInboxTest() {
-        when(dataSafeService.inboxService().read(any())).thenReturn(new ByteArrayInputStream("hello".getBytes()));
-
-        mvc.perform(get("/inbox/{path}", TEST_PATH)
+        String path = "path/to/file";
+        mvc.perform(get("/document/{path}", path)
                 .header("user", TEST_USER)
                 .header("password", TEST_PASS)
                 .accept(MediaType.APPLICATION_OCTET_STREAM_VALUE))
@@ -73,9 +61,32 @@ public class InboxControllerTest {
 
     @SneakyThrows
     @Test
-    public void deleteFromInboxTest() {
+    public void writeDocumentTest() {
+        when(dataSafeService.privateService().write(any())).thenReturn(new ByteArrayOutputStream());
+        String path = "path/to/file";
+        mvc.perform(put("/document/{path}", path)
+                .contentType(MediaType.APPLICATION_OCTET_STREAM_VALUE)
+                .header("user", TEST_USER)
+                .header("password", TEST_PASS)
+        )
+                .andExpect(status().isOk());
+    }
 
-        mvc.perform(delete("/inbox/{path}", TEST_PATH)
+    @SneakyThrows
+    @Test
+    public void listDocumentsTest() {
+        String path = "";
+        mvc.perform(get("/documents/{path}", path)
+                .header("user", TEST_USER)
+                .header("password", TEST_PASS)
+        ).andExpect(status().isOk());
+    }
+
+    @SneakyThrows
+    @Test
+    public void deleteDocumentTest() {
+        String path = "path/to/file";
+        mvc.perform(delete("/document/{path}", path)
                 .header("user", TEST_USER)
                 .header("password", TEST_PASS)
         ).andExpect(status().isOk());
