--- datasafe-rest-impl/src/test/java/de/adorsys/datasafe/rest/impl.java
@@ -1,4 +1,0 @@
-package de.adorsys.datasafe.rest;
-
-public class impl {
-}
